+++ src/connectedk8s/azext_connectedk8s/_constants.py
@@ -65,3 +65,7 @@ Helm_Existing_User_Supplied_Value_Get_Fault = 'Error while loading the user supp
 Error_Flattening_User_Supplied_Value_Dict = 'Error while flattening the user supplied helm values dict'
 Upgrade_RG_Cluster_Name_Conflict = 'The provided cluster name and rg correspond to different cluster'
 Corresponding_CC_Resource_Deleted_Fault = 'CC resource corresponding to this cluster has been deleted by the customer'
+Enable_Azure_RBAC_Not_Supported_Fault_Type = 'enable-azure-rbac-not-supported-error'
+Enable_Azure_RBAC_Not_Supported_Error = 'Enabling Azure RBAC is supported for versions less that 1.17'
+Insufficient_Args_Fault_Type = 'insufficient-args-error'
+Insufficient_Args_Fault_Error = 'Please specify {} when using {}'