+++ src/connectedk8s/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.2.10
+++++++
+* `az connectedk8s connect`: Added support for enabling azure rbac 
+* `az connectedk8s update`: Added support for updating whether to use azure rbac 
 
 0.2.9
 ++++++